+++ packages/cli/src/lib/commands/collect/processes/collect-reports.ts
@@ -1,10 +1,10 @@
 import { RcJson } from '../../../types';
 import { UserFlowProvider } from '../utils/user-flow/types';
-import { collectFlow, loadFlow, openFlowReport, persistFlow } from '../utils/user-flow';
 import { concat } from '../../../core/utils/processing/behaviors';
 import { logVerbose } from '../../../core/utils/loggin';
 import { get as dryRun } from '../../../core/options/dryRun';
 import { AssertOptions } from '../../../core/rc-json/types';
+import { collectFlow, openFlowReport, persistFlow, loadFlow } from '../utils/user-flow';
 
 
 export async function collectReports(cfg: RcJson): Promise<RcJson> {
