+++ packages/cli/src/lib/index.ts
@@ -1,10 +1,7 @@
 export * from './types';
 export { UserFlowMock } from './commands/collect/utils/user-flow/user-flow.mock';
-export { StepOptions } from './commands/collect/utils/user-flow/types';
-export { UserFlowOptions } from './commands/collect/utils/user-flow/types';
-export { LaunchOptions } from './commands/collect/utils/user-flow/types';
-export { UserFlowContext } from './commands/collect/utils/user-flow/types';
-export { UserFlowProvider } from './commands/collect/utils/user-flow/types';
-export { UserFlowInteractionsFn } from './commands/collect/utils/user-flow/types';
+export {
+  StepOptions, UserFlowOptions, LaunchOptions, UserFlowContext, UserFlowProvider, UserFlowInteractionsFn
+} from './commands/collect/utils/user-flow/types';
 export { createUserFlowRunner } from './commands/collect/utils/replay';
 export { MeasureModes } from './commands/collect/utils/replay/types';
